ref(core): Move `clack-utils.ts` to dedicated folder
