+++ src/nuxt/utils.ts
@@ -3,7 +3,7 @@ import * as clack from '@clack/prompts';
 import { gte, minVersion } from 'semver';
 // @ts-expect-error - magicast is ESM and TS complains about that. It works though
 import { loadFile } from 'magicast';
-import { abortIfCancelled } from '../utils/clack-utils';
+import { abortIfCancelled } from '../utils/clack';
 
 export async function isNuxtV4(
   nuxtConfig: string,
