+++ src/sourcemaps/tools/create-react-app.ts
@@ -1,6 +1,6 @@
 // @ts-expect-error - clack is ESM and TS complains about that. It works though
 import clack from '@clack/prompts';
-import { abortIfCancelled } from '../../utils/clack-utils';
+import { abortIfCancelled } from '../../utils/clack';
 
 export async function configureCRASourcemapGenerationFlow(): Promise<void> {
   await abortIfCancelled(
