+++ src/sourcemaps/tools/remix.ts
@@ -3,7 +3,7 @@ import * as clack from '@clack/prompts';
 import chalk from 'chalk';
 import { runRemixWizard } from '../../remix/remix-wizard';
 import { traceStep } from '../../telemetry';
-import { abortIfCancelled } from '../../utils/clack-utils';
+import { abortIfCancelled } from '../../utils/clack';
 import type { WizardOptions } from '../../utils/types';
 import type { SourceMapUploadToolConfigurationOptions } from './types';
