+++ src/sourcemaps/utils/other-wizards.ts
@@ -3,11 +3,7 @@ import clack from '@clack/prompts';
 import chalk from 'chalk';
 import { runSvelteKitWizard } from '../../sveltekit/sveltekit-wizard';
 
-import {
-  abort,
-  abortIfCancelled,
-  getPackageDotJson,
-} from '../../utils/clack-utils';
+import { abort, abortIfCancelled, getPackageDotJson } from '../../utils/clack';
 import {
   findInstalledPackageFromList,
   hasPackageInstalled,
